@@ -1,8 +1,5 @@
 /*!
  * Copyright (c) 2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
 
-const issuerName = process.env.ISSUER_NAME || 'Digital Bazaar';
-
-module.exports = {issuerName};
+export const issuerName = process.env.ISSUER_NAME || 'Digital Bazaar';
